pyproject.toml
[tool.poetry]
name = "track-bitcoin"
version = "0.1.0"
description = "Bitcoin price track"
authors = ["nilton seixas <nfsseixas@gmail.com>"]
readme = "README.md"
packages = [{include = "track_bitcoin"}]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.28.2"
yfinance = "^0.2.14"
python-telegram-bot = "^20.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
